Refactor Waiver module to use Waiver.Store
* Move Repo actions to Waiver.Store
* Closes #285
