+++ lib/ex338_web/controllers/waiver_controller.ex
@@ -1,7 +1,7 @@
 defmodule Ex338Web.WaiverController do
   use Ex338Web, :controller
 
-  alias Ex338.{FantasyLeague, FantasyTeam, FantasyPlayer, Waiver}
+  alias Ex338.{FantasyLeague, FantasyTeam, FantasyPlayer, Waiver, Waiver.Store}
   alias Ex338Web.{Authorization, NotificationEmail}
 
   import Canary.Plugs
@@ -19,7 +19,7 @@ defmodule Ex338Web.WaiverController do
   def index(conn, %{"fantasy_league_id" => league_id}) do
     render(conn, "index.html",
       fantasy_league: FantasyLeague.Store.get(league_id),
-      waivers:        Waiver.get_all_waivers(league_id)
+      waivers:        Store.get_all_waivers(league_id)
     )
   end
 
@@ -39,7 +39,7 @@ defmodule Ex338Web.WaiverController do
   def create(conn, %{"fantasy_team_id" => _id, "waiver" => waiver_params}) do
     team = conn.assigns.fantasy_team
 
-    case Waiver.create_waiver(team, waiver_params) do
+    case Store.create_waiver(team, waiver_params) do
       {:ok, waiver} ->
         NotificationEmail.waiver_submitted(waiver)
 
@@ -74,7 +74,7 @@ defmodule Ex338Web.WaiverController do
   def update(conn, %{"id" => _id, "waiver" => params}) do
     waiver = conn.assigns.waiver
 
-    case Waiver.update_waiver(waiver, params) do
+    case Store.update_waiver(waiver, params) do
       {:ok, waiver} ->
         conn
         |> put_flash(:info, "Waiver successfully updated")
